Issue-456: Movie module created
