+++ app/src/main/java/com/kirchhoff/movies/di/NetworkModule.kt
@@ -3,8 +3,7 @@ package com.kirchhoff.movies.di
 import com.facebook.stetho.okhttp3.StethoInterceptor
 import com.kirchhoff.movies.BuildConfig
 import com.kirchhoff.movies.network.interceptors.ApiKeyInterceptor
-import com.kirchhoff.movies.network.services.DiscoverService
-import com.kirchhoff.movies.network.services.MovieService
+import com.kirchhoff.movies.screen.movie.network.MovieService
 import com.kirchhoff.movies.screen.person.network.PersonService
 import com.kirchhoff.movies.screen.tvshow.network.TvShowService
 import okhttp3.OkHttpClient
@@ -27,8 +26,6 @@ val networkModule = module {
             .addConverterFactory(GsonConverterFactory.create())
             .build()
     }
-
-    single { get<Retrofit>().create(DiscoverService::class.java) }
 
     single { get<Retrofit>().create(PersonService::class.java) }
