+++ app/src/main/java/com/kirchhoff/movies/di/RepositoryModule.kt
@@ -1,9 +1,7 @@
 package com.kirchhoff.movies.di
 
-import com.kirchhoff.movies.repository.discover.DiscoverRepository
-import com.kirchhoff.movies.repository.discover.IDiscoverRepository
-import com.kirchhoff.movies.repository.movie.IMovieRepository
-import com.kirchhoff.movies.repository.movie.MovieRepository
+import com.kirchhoff.movies.screen.movie.repository.IMovieRepository
+import com.kirchhoff.movies.screen.movie.repository.MovieRepository
 import com.kirchhoff.movies.screen.person.repository.IPersonsRepository
 import com.kirchhoff.movies.screen.person.repository.PersonsRepository
 import com.kirchhoff.movies.screen.tvshow.repository.ITvShowRepository
@@ -11,12 +9,6 @@ import com.kirchhoff.movies.screen.tvshow.repository.TvShowRepository
 import org.koin.dsl.module
 
 val repositoryModule = module {
-    single<IDiscoverRepository> {
-        DiscoverRepository(
-            discoverService = get(),
-            discoverMapper = get()
-        )
-    }
     single<IPersonsRepository> {
         PersonsRepository(
             personService = get(),
@@ -27,7 +19,8 @@ val repositoryModule = module {
     single<IMovieRepository> {
         MovieRepository(
             movieService = get(),
-            movieDetailsMapper = get()
+            movieDetailsMapper = get(),
+            discoverMapper = get()
         )
     }
     single<ITvShowRepository> {
