--- app/src/main/java/com/kirchhoff/movies/repository/discover/DiscoverRepository.kt
@@ -1,19 +1,0 @@
-package com.kirchhoff.movies.repository.discover
-
-import com.kirchhoff.movies.core.data.UIMovie
-import com.kirchhoff.movies.core.mapper.IDiscoverMapper
-import com.kirchhoff.movies.core.repository.BaseRepository
-import com.kirchhoff.movies.core.repository.Result
-import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.network.services.DiscoverService
-
-class DiscoverRepository(
-    private val discoverService: DiscoverService,
-    private val discoverMapper: IDiscoverMapper
-) : BaseRepository(), IDiscoverRepository {
-
-    override suspend fun fetchMovies(page: Int): Result<UIPaginated<UIMovie>> =
-        discoverMapper.createUIDiscoverMovieList(apiCall {
-            discoverService.fetchDiscoverMovie(page)
-        })
-}
