--- app/src/main/java/com/kirchhoff/movies/repository/discover/IDiscoverRepository.kt
@@ -1,9 +1,0 @@
-package com.kirchhoff.movies.repository.discover
-
-import com.kirchhoff.movies.core.data.UIMovie
-import com.kirchhoff.movies.core.repository.Result
-import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-
-interface IDiscoverRepository {
-    suspend fun fetchMovies(page: Int): Result<UIPaginated<UIMovie>>
-}
