+++ app/src/main/res/values/strings.xml
@@ -3,13 +3,4 @@
     <string name="menu_movies">Movies</string>
     <string name="menu_tv">Tvs</string>
     <string name="menu_persons">Persons</string>
-    <string name="retry">Retry</string>
-    <string name="trailers">Trailers</string>
-
-    <string name="reviews">Reviews</string>
-    <string name="similar_movies">Similar Movies</string>
-
-    <string name="movie_runtime_format">%1$s / %2$d min</string>
-
-    <string name="empty_movies">Can\'t load movies</string>
 </resources>
