+++ screen/movie/src/main/java/com/kirchhoff/movies/screen/movie/mapper/details/IMovieDetailsMapper.kt
@@ -1,0 +1,15 @@
+package com.kirchhoff.movies.screen.movie.mapper.details
+
+import com.kirchhoff.movies.core.data.UIEntertainmentCredits
+import com.kirchhoff.movies.core.repository.Result
+import com.kirchhoff.movies.networkdata.core.NetworkEntertainmentCredits
+import com.kirchhoff.movies.networkdata.details.movie.NetworkMovieDetails
+import com.kirchhoff.movies.networkdata.details.movie.NetworkTrailersList
+import com.kirchhoff.movies.screen.movie.data.UIMovieDetails
+import com.kirchhoff.movies.screen.movie.data.UITrailersList
+
+interface IMovieDetailsMapper {
+    fun createUIMovieDetails(movieDetailsResult: Result<NetworkMovieDetails>): Result<UIMovieDetails>
+    fun createUIEntertainmentCredits(movieCreditsResult: Result<NetworkEntertainmentCredits>): Result<UIEntertainmentCredits>
+    fun createUITrailersList(trailersListResult: Result<NetworkTrailersList>): Result<UITrailersList>
+}
