+++ screen/movie/src/main/res/layout/fragment_movie_details.xml
