+++ screen/movie/src/main/res/layout/item_movie.xml
