+++ settings.gradle
@@ -5,6 +5,7 @@ include ':app',
         ':view:keywordsview',
         ':view:creditsview',
         ':view:ratingtextview',
+        ':screen:movie',
         ':screen:tvshow',
         ':screen:review',
         ':screen:country',
